requirements.txt
python-telegram-bot==20.8
python-dotenv  # Эта библиотека не нужна для Render, но полезна локально для тестирования токена